Fix faux-bold on headlines
Currently, only the Medium weight of Fira Sans is being loaded from
Google Fonts, but the font-weight property of the main headline and
sub-headings are set to weights that correspond to heavier styles. In
most browsers this is fine, but in Safari on mobile and desktop, it
results in a subtle faux-bold effect.

This change updates the font-weight for h1, h2, h3 and strong in
style.scss.